feat: 扩展内核全面重构、gamehost启动增强、移动端连接与测试完善
Backend:
- extension/kernel: capability系统重构（新增extension_provider_reconciler/provider_instance_reconciler/provider_invocation/provider_operation/resolution/resolution_policy/resolver/service），builtin模块迁移，tool_facade/task_runtime新增capability_placement_resolver，execution新增approval_gate，manifest_v2全面重构，permission新增sqlite_snapshot_store，contribution_installer/package_install_saga升级
- gamehost: startup模块（adapters/coordinator），compose/control/default_effect_sink/control_effect_sink升级
- androidnative/interaction: 更新models
- desktoppet: 新增production_capabilities，更新implementations
- server: 更新runtime_bootstrap_other/services

Mobile:
- 全面更新backend_connection（repository/source/providers/runtime_backend_connection_source）
- 新增default_runtime_status_projection
- iOS: 更新BluetoothNativeHandler/HomeKitNativeHandler
- Flutter: 更新app.dart
- 新增大量测试覆盖（gate_projection/backend_websocket/runtime_status/desktop_pet等）

Tests:
- gamehost_external: 新增backend_driver/game_center_client/waiters，更新harNESS/matrices全部测试

+++ backend/internal/gamehost/integration/control_effect_sink.go
@@ -72,7 +72,7 @@ func (s *ProtocolControlEffectSink) ExecuteAuthorized(
 		RuntimeID:  runtimeID,
 		PluginID:   pluginID,
 		ServiceID:  serviceID,
-		Generation: int(permit.Generation),
+		Generation: int64(permit.Generation),
 	}
 
 	envelope := protocol.Envelope{
